+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## 1.0.0-beta2 - xx-xxx-2013
 
 - `$sluggable` property of model switched to static.
+- update type hinting in Sluggable::make()
 
 
 ## 1.0.0-beta - 11-Jun-2013